Updated by Kamalakar Dandu

+++ partyanalyst/src/main/java/com/itgrids/partyanalyst/web/action/VotersEditAction.java
@@ -1760,7 +1760,7 @@ public String saveLocality()
 	   public String updateMobileNos()
 	   {
 		   try{
-			   result = voterReportService.updateMobileNos(getStartIndex());	   
+			   result = voterReportService.updateMobileNos(getStartIndex().longValue());	   
 		   }
 		   catch(Exception e)
 		   {
